Remove usage of `common.Amount` in type structs

Closes #94

See merge request thorchain/bepswap/thornode!250

+++ x/thorchain/mocks/mock_pool_storage.go
@@ -71,6 +71,6 @@ func (mps MockPoolStorage) GetAdminConfigStakerAmtInterval(ctx sdk.Context, addr
 
 func (mps MockPoolStorage) AddIncompleteEvents(ctx sdk.Context, event types.Event) {}
 
-func (mps MockPoolStorage) GetLowestActiveVersion(ctx sdk.Context) int {
+func (mps MockPoolStorage) GetLowestActiveVersion(ctx sdk.Context) int64 {
 	return 0
 }
